wl_trader_re_v3_3_0/server/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-socketio==5.11.3
ccxt==4.3.92
pyzmq==26.2.0
SQLAlchemy==2.0.35
alembic==1.14.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
